IDE0004: Remove unnecessary cast

+++ src/DapperRepository.BulkInsert.cs
@@ -54,7 +54,7 @@ public partial class DapperRepository<TEntity>
                 var maxAllowedInstancesPerBatch = totalInstances / exceededTimes;
                 if (maxAllowedInstancesPerBatch > 1000) maxAllowedInstancesPerBatch = 1000;
 
-                var maxIterationCount = (int)Math.Ceiling((double)totalInstances / (double)maxAllowedInstancesPerBatch);
+                var maxIterationCount = (int)Math.Ceiling(totalInstances / (double)maxAllowedInstancesPerBatch);
 
                 for (var i = 0; i <= maxIterationCount; i++)
                 {
@@ -101,7 +101,7 @@ public partial class DapperRepository<TEntity>
                 var maxAllowedInstancesPerBatch = totalInstances / exceededTimes;
                 if (maxAllowedInstancesPerBatch > 1000) maxAllowedInstancesPerBatch = 1000;
 
-                var maxIterationCount = (int)Math.Ceiling((double)totalInstances / (double)maxAllowedInstancesPerBatch);
+                var maxIterationCount = (int)Math.Ceiling(totalInstances / (double)maxAllowedInstancesPerBatch);
 
                 for (var i = 0; i <= maxIterationCount; i++)
                 {
